- Terminadas interfaces visuales de las ventanas de profesores y alumnos.
