+++ sistema de gestion/src/Profesor/IngresarProfesor.java
@@ -1,6 +1,8 @@
 
 package Profesor;
 
+import Ventanas.TipoVentana.ETipoVentana;
+
 /**
  *
  * @author Varela Vargas Leandro Gastón
@@ -11,7 +13,7 @@ public class IngresarProfesor extends ModeloVentanaProfesor
     public IngresarProfesor ()
     {
         
-        super ( "Ingresar Docente", ModeloVentanaProfesor.TipoVentana.Crear );
+        super ( "Ingresar Docente", ETipoVentana.Crear );
         
     }
