+++ sistema de gestion/src/Profesor/ModificarProfesor.java
@@ -1,6 +1,8 @@
 
 package Profesor;
 
+import Ventanas.TipoVentana.ETipoVentana;
+
 /**
  *
  * @author Varela Vargas Leandro Gastón
@@ -10,7 +12,7 @@ public class ModificarProfesor extends ModeloVentanaProfesor
  
     public ModificarProfesor ()
     {
-        super ( "Modificar Docente", ModeloVentanaProfesor.TipoVentana.Modificar );
+        super ( "Modificar Docente", ETipoVentana.Modificar );
     }
     
 }
