+++ sistema de gestion/src/Ventanas/TipoVentana.java
@@ -1,0 +1,15 @@
+
+package Ventanas;
+
+/**
+ *
+ * @author VarelaVargasLeandro
+ */
+public class TipoVentana
+{
+    public static enum ETipoVentana
+    {
+        Crear,
+        Modificar
+    }
+}
